* 오타 수정 * 메모리 트레이서 함수 정리 * 코어의 스레드 함수 분리
